fix: change datatype location coordinate, add converter for bigdecimal

+++ src/main/java/org/nefedov/weather/application/dto/CoordinateDto.java
@@ -1,4 +1,6 @@
 package org.nefedov.weather.application.dto;
 
-public record CoordinateDto(Double latitude, Double longitude) {
+import java.math.BigDecimal;
+
+public record CoordinateDto(BigDecimal latitude, BigDecimal longitude) {
 }
